libta: drop lib prefix, rename to 'ta'
