+++ src/CMakeLists.txt
@@ -1,2 +1,2 @@
-add_subdirectory(libta)
+add_subdirectory(ta)
 add_subdirectory(mtl)
