+++ src/ta/ata.cpp
@@ -18,8 +18,8 @@
  *  Read the full text in the LICENSE.GPL file in the doc directory.
  */
 
-#include <libta/ata.h>
-#include <libta/automata.h>
+#include <ta/ata.h>
+#include <ta/automata.h>
 
 #include <cassert>
 #include <iterator>
