+++ src/ta/include/ta/automata.h
