+++ test/CMakeLists.txt
@@ -3,7 +3,7 @@ find_package(Catch2 REQUIRED)
 include(Catch)
 
 add_executable(testta test_ta.cpp catch2_main.cpp)
-target_link_libraries(testta PRIVATE libta PRIVATE Catch2::Catch2)
+target_link_libraries(testta PRIVATE ta PRIVATE Catch2::Catch2)
 catch_discover_tests(testta)
 
 add_executable(testinterval test_interval.cpp catch2_main.cpp)
@@ -15,5 +15,5 @@ target_link_libraries(testmtlformulae PRIVATE mtl PRIVATE Catch2::Catch2)
 catch_discover_tests(testmtlformulae)
 
 add_executable(testata catch2_main.cpp test_ata_formula.cpp test_ata.cpp)
-target_link_libraries(testata PRIVATE libta PRIVATE Catch2::Catch2)
+target_link_libraries(testata PRIVATE ta PRIVATE Catch2::Catch2)
 catch_discover_tests(testata)
